[ALPS03574894] GED: use u64 instead of unsigned long long for BBQ_ID
[Detail]
use u64 instead of unsigned long long for BBQ_ID

MTK-Commit-Id: c9f3184992a99ff1f1ef54465281bdaae8324f3b

CR-Id: ALPS03574894
Feature: GPU DVFS

+++ drivers/misc/mediatek/gpu/ged/include/ged_kpi.h
@@ -21,16 +21,16 @@
 #include <linux/topology.h>
 
 GED_ERROR ged_kpi_dequeue_buffer_ts(int pid,
-								unsigned long long ullWdnd,
+								u64 ullWdnd,
 								int i32FrameID,
 								int fence_fd,
 								int isSF);
 GED_ERROR ged_kpi_queue_buffer_ts(int pid,
-								unsigned long long ullWdnd,
+								u64 ullWdnd,
 								int i32FrameID,
 								int fence,
 								int QedBuffer_length);
-GED_ERROR ged_kpi_acquire_buffer_ts(int pid, unsigned long long ullWdnd, int i32FrameID);
+GED_ERROR ged_kpi_acquire_buffer_ts(int pid, u64 ullWdnd, int i32FrameID);
 GED_ERROR ged_kpi_sw_vsync(void);
 GED_ERROR ged_kpi_hw_vsync(void);
 int ged_kpi_get_uncompleted_count(void);
